rename back `errors.ts`, merge `errmap.ts` into `_errors.ts`
